requirements.txt
datasketch
numpy
scikit-learn
umap-learn
matplotlib
mido
python-rtmidi
